updated clm

+++ Coelengine/src/Coel/Graphics/Renderer.hpp
@@ -10,7 +10,7 @@ namespace Coel {
 			void flush();
 
 			void drawRect(float x, float y, float w, float h, unsigned int color = -1);
-			void resizeViewport(float x, float y, float w, float h);
+			void resizeViewport(int x, int y, int w, int h);
 		} // namespace Renderer
 	} // namespace Graphics
 } // namespace Coel
